+++ inc/actions.php
@@ -9,9 +9,9 @@
  */
 add_action('init', 'uploadcare_plugin_init');
 function uploadcare_plugin_init() {
-    wp_register_script(
-        'uploadcare-widget',
-        'https://ucarecdn.com/widget/0.17.2/uploadcare/uploadcare-0.17.2.min.js');
+    $widget_url = sprintf('https://ucarecdn.com/widget/%s/uploadcare/uploadcare-%s.min.js',
+        UPLOADCARE_WIDGET_VERSION, UPLOADCARE_WIDGET_VERSION);
+    wp_register_script('uploadcare-widget', $widget_url);
 
     wp_register_script(
         'uploadcare-config',
@@ -30,7 +30,7 @@ function uploadcare_plugin_init() {
         array('uploadcare-config'));
 
 
-    // _uploadcare_register_user_images();
+    _uploadcare_register_user_images();
 }
 
 
@@ -56,7 +56,7 @@ add_action('media_buttons_context', 'uploadcare_add_media');
 function uploadcare_add_media($context) {
     $api = uploadcare_api();
 
-    $img = plugins_url('logo.png', __FILE__);
+    $img = plugins_url('media/logo.png', dirname(__FILE__));
 
     if(get_option('uploadcare_finetuning')) {
         $finetuning = stripcslashes(get_option('uploadcare_finetuning'));
@@ -110,6 +110,7 @@ function uploadcare_attach($file) {
     add_post_meta($attachment_id, '_wp_attached_file', $file->data['original_file_url'], true);
     add_post_meta($attachment_id, '_wp_attachment_metadata', $meta, true);
     add_post_meta($attachment_id, 'uploadcare_url', $file->data['original_file_url'], true);
+    return $attachment_id;
 }
 
 /**
@@ -124,6 +125,52 @@ function uploadcare_handle() {
     $file->store();
 
     uploadcare_attach($file);
+}
+
+
+/**
+ * Create User Image post
+ *
+ * @param $file Uploadcare File object to attach
+ * @param $post_id Post ID image should be attached to
+ */
+function uploadcare_attach_user_image($file, $post_id) {
+    $attachment_id = uploadcare_attach($file);
+    $currentuser = get_current_user_id();
+
+
+    $filename = $file->data['original_filename'];
+    $title = $filename;
+
+    $user_image = array(
+        'post_author'    => $currentuser,
+        'post_date'      => date('Y-m-d H:i:s'),
+        'post_type'      => 'uc_user_image',
+        'post_title'     => $title,
+        'post_parent'    => $post_id,
+        'post_status'    => 'private',
+        'post_mime_type' => $file->data['mime_type'],
+    );
+
+    $user_image_id = wp_insert_post($user_image, true);
+    set_post_thumbnail($user_image_id, $attachment_id);
+}
+
+/**
+ * Add ajax handler for uploadcare shortcode
+ */
+add_action('wp_ajax_uploadcare_shortcode_handle', 'uploadcare_shortcode_handle');
+add_action('wp_ajax_nopriv_uploadcare_shortcode_handle', 'uploadcare_shortcode_handle');
+function uploadcare_shortcode_handle() {
+    // store file
+    $api = uploadcare_api();
+    $file_id = $_POST['file_id'];
+    $post_id = $_POST['post_id'];
+    $file = $api->getFile($file_id);
+    $file->store();
+
+    // create user image
+    uploadcare_attach_user_image($file, $post_id);
 }
 
 
@@ -145,120 +192,6 @@ function uploadcare_media_files_menu_handle() {
 }
 
 
-/**
- * Remove Featured Image Meta Box to be replaced with Uploadcare featured image box
- */
-add_action('do_meta_boxes', 'uploadcare_remove_wp_featured_image_box');
-function uploadcare_remove_wp_featured_image_box() {
-    if (get_option('uploadcare_replace_featured_image')) {
-        remove_meta_box('postimagediv', NULL, 'side');
-    }
-}
-
-
-/**
- * Prints Uploadcare featured image box content.
- *
- * @param WP_Post $post The object for the current post/page.
- */
-function uploadcare_featured_image_box($post) {
-    // Add an nonce field so we can check for it later.
-    wp_nonce_field('uploadcare_featured_image_box',
-                 'uploadcare_featured_image_box_nonce');
-
-    $value = get_post_meta($post->ID, 'uploadcare_featured_image', true);
-    $html = <<<HTML
-<a title="Set featured image"
-   id="uc-set-featured-img"
-   href="javascript:;"
-   data-uc-url="{$value}">Set featured image</a>
-<a title="Remove featured image"
-   id="uc-remove-featured-img"
-   href="javascript:;"
-   class="hidden">Remove featured image</a>
-<input type="hidden"
-       id="uc-featured-image-input"
-       name="uploadcare_featured_image"
-       value="{$value}">
-HTML;
-    echo $html;
-}
-
-
-/**
- * Adds a box to the main column on the Post and Page edit screens.
- */
-add_action('add_meta_boxes', 'uploadcare_add_featured_image_box');
-function uploadcare_add_featured_image_box($post_type) {
-  if (get_option('uploadcare_replace_featured_image') &&
-      post_type_supports($post_type, 'thumbnail')) {
-
-      add_meta_box(
-          'myplugin_sectionid',
-          __('Featured Image (uploadcare)', 'uploadcare'),
-          'uploadcare_featured_image_box',
-          $post_type,
-          'side'
-      );
-  }
-}
-
-
-/**
- * When the post is saved, saves our custom data.
- *
- * @param int $post_id The ID of the post being saved.
- */
-add_action('save_post', 'uploadcare_save_postdata');
-function uploadcare_save_postdata($post_id) {
-    // at the moment this is used only for featured images, so skip it if
-    //   the option is not set
-    if (!get_option('uploadcare_replace_featured_image')) {
-        return $post_id;
-    }
-
-    /*
-    * We need to verify this came from the our screen and with proper authorization,
-    * because save_post can be triggered at other times.
-    */
-
-    // Check if our nonce is set.
-    if (!isset( $_POST['uploadcare_featured_image_box_nonce'])) {
-        return $post_id;
-    }
-    $nonce = $_POST['uploadcare_featured_image_box_nonce'];
-
-    // Verify that the nonce is valid.
-    if (!wp_verify_nonce($nonce, 'uploadcare_featured_image_box')) {
-        return $post_id;
-    }
-
-    // If this is an autosave, our form has not been submitted, so we don't want to do anything.
-    if (defined('DOING_AUTOSAVE') && DOING_AUTOSAVE) {
-        return $post_id;
-    }
-
-    // Check the user's permissions.
-    if ('page' == $_POST['post_type']) {
-        if (!current_user_can('edit_page', $post_id)) {
-            return $post_id;
-        }
-    } else {
-        if (!current_user_can('edit_post', $post_id)) {
-            return $post_id;
-        }
-    }
-
-    /* OK, its safe for us to save the data now. */
-
-    // Sanitize user input.
-    $mydata = sanitize_text_field($_POST['uploadcare_featured_image']);
-
-    // Update the meta field in the database.
-    update_post_meta($post_id, 'uploadcare_featured_image', $mydata);
-}
-
-
 // add_action('post-upload-ui', 'uploadcare_media_upload');
 function uploadcare_media_upload() {
 
@@ -267,7 +200,7 @@ function uploadcare_media_upload() {
     // elseif ( ! empty( $GLOBALS['post_ID'] ) )
     //     $browser_uploader .= '&amp;post_id=' . intval( $GLOBALS['post_ID'] );
 
-    $img = plugins_url('logo.png', __FILE__);
+    $img = plugins_url('media/logo.png', dirname(__FILE__));
 
     ?>
 
@@ -299,7 +232,7 @@ function uploadcare_display_thumbnail_column($col, $id) {
 
 
 function uploadcare_settings() {
-    include('inc/uploadcare_settings.php');
+    include('uploadcare_settings.php');
 }