+++ inc/filters.php
@@ -37,6 +37,7 @@ function uploadcare_image_downsize($value = false, $id, $size = 'medium') {
  * - use uploadcare for featured images is set
  * - post's meta 'uploadcare_featured_image' is set
  * otherwise, uses default html code.
+ * @legacy
  */
 add_filter('post_thumbnail_html', 'uploadcare_post_thumbnail_html', 10, 5);
 function uploadcare_post_thumbnail_html($html, $post_id, $post_thumbnail_id, $size, $attr) {